Added GCAsset Class and GCThumbnailComponent class.  Made initWithDictionary method of GCResource class accessable.
GCThumbnailComponent has a default nib file, but can be subclassed and
use either a custom nib, or have the UI setup programatically.  Until
Assets are fixed on staging I have a blue and yellow background on
images so you can tell that they are there.  Once images are working
properly again I will change them to clear or white.