Finished GCAssetSliderComponent, Set up an account on dev server so i could test components easier.  Adjusted GCThumbnailComponent to so that it looked better.
Finished GCAssetSliderComponent.  It has a default nib file, but can be
subclassed and use a custom nib or programmatically created UI.  Images
in slider are zoomable up to 2.5x.  I can make this adjustable in a
future iteration.  Adjusted full size images in GCThumbnailComponent to
scale to fit and have a black background.