Several Slight adjustments to various odes and components.
Changed GCImageSelectionComponent to inherit from
GCUIBaseViewController. Changed GCAsset to always use alAsset for
thumbnail if it has one regardless of the asset's status. Adjusted
GCThumbnailComponent to show loading indicator when loading the full
screen image since the first call for a size from the media server can
take a little while. Adjusted uploading in GCParcel to account for the
orientation of the image so that images don't show up sideways.