Adjusted UI of GCUploadComponent.
Adjusted the look of the default slider on the bottom of the
GCUploadComponent nib.  Also made it so that the slider portion is
completely optional.  If a subclass doesn't put it in either in code or
in the nib file then it will simply be skipped over and all other
functionality of the class will remain intact.