Fixed error in GCAccount class, added method to retrieve creation date of assets, add component to generate a pre made email editor for all attendees of an event.
Created GCEventEmailComponent, which has a class method that returns a
MFMailViewController that is initialized with the email addresses of
attendees of an event.  You supply the event name and an optional date
range.  If no range is supplied it looks thirty days ahead and behind
the current day.  You can also supply an optional subject and message
for it to initialize the controller with.  Imported GCChute.h in
GCAccount to suppress error.  Added method to retrieve date a GCAsset
was created for use in the upcoming filtered asset picker.  Date will
only work with GCAssets that have an ALAsset as assets from server do
not have a date parameter.  Assets from server will return NULL.